% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.list.R
\name{coef.list}
\alias{coef.list}
\title{Extract Model Coefficients}
\usage{
\method{coef}{list}(model)
}
\arguments{
\item{model}{A \code{list} as computed by \code{\link{reduction}}.}
}
\description{
Extract Model Coefficients
}
