% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\alias{conf_int.list}
\alias{conf_int.lm}
\title{Confidence intervals for model covariates}
\usage{
conf_int(model, ...)

\method{conf_int}{lm}(model, covariates, level = 0.95)

\method{conf_int}{list}(model, covariates, level = 0.95)
}
\arguments{
\item{model}{A fitted model.}

\item{covariates}{A character vector giving the names of the covariates in 
the model. When missing, the confidence intervals are computed for all the
covariates.}

\item{level}{The confidence level required. Default is 95\%.}
}
\value{
A matrix (or vector) with columns giving the estimated coefficients,
the lower and upper confidence limits for each parameter.
}
\description{
Computes confidence intervals for one or more covariates in a fitted model.
}
\details{
This function computes confidence intervals without calling 
\code{\link[stats]{vcov}}, but it still assumes normality. This brings slightly
different results with respect to the standar 
\code{\link[stats]{confint}} method. Either a linear 
model given by \code{\link[stats]{lm}} or a reduced model given by 
\code{\link{reduction}} can be provided.
}
\examples{
model = lm( mpg ~ hp + wt, data = mtcars )

# compute confidence intervals for all covariates
conf_int( model )

# specify for which covariate a confidence intervals must be computed
conf_int( model, "wt" )
}
\author{
Francesco Grossetti \email{francesco.grossetti@gmail.com}.
}
\seealso{
\code{\link[stats]{lm}} \code{\link[statutils]{reduction}} 
\code{\link[stats]{confint}}
}

