% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_purpose.R
\name{count_na}
\alias{count_na}
\title{Utility to count missing data in a \code{matrix}, \code{data.frame} or 
\code{data.table}.}
\usage{
count_na(data, prop = TRUE)
}
\arguments{
\item{data}{A \code{matrix}, \code{data.frame} or \code{data.table}.}

\item{prop}{Compute the column proportion of missing data.}
}
\value{
Print the results in console.
}
\description{
The function returns the sum of missing data detected in each column.
}
\author{
Francesco Grossetti \email{francesco.grossetti@gmail.com}.
}
