% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{def_complete}
\alias{def_complete}
\title{Complete a dataset after \code{\link[mice]{mice}} has been run.}
\usage{
def_complete(data, data_key)
}
\arguments{
\item{data}{A \code{data.table} or a \code{data.frame} (see 'Details').}

\item{data_key}{The keying variable in the \code{data.table}.}
}
\value{
Either a \code{data.table} or a \code{data.frame}.
}
\description{
Utility to define a non-repeated dataset after a missing imputation algorithm 
with the function \code{\link[mice]{mice}} has been run and the associated 
complete procedure through \code{\link[mice]{complete}} has been carried out.
}
\details{
The function acts on numeric variables only and takes a column mean
by considering all the simulations detected given by \code{.imp}. This should
ensure consistent results as given by \code{\link[mice]{pool}}
When a \code{data.frame} is passed, \code{def_complete} internally
converts it to a \code{data.table} but still returns a \code{data.frame}.
}
\author{
Francesco Grossetti \email{francesco.grossetti@gmail.com}.
}

