% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduction}
\alias{reduction}
\alias{reduction.lm}
\title{Reduce a linear model by iteratively checking the signficance of each predictor.}
\usage{
reduction(model, ...)

\method{reduction}{lm}(model, data, alpha = 0.05, intercept = TRUE)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{data}{A \code{data.frame} or \code{data.table} on which \code{model} has run.}

\item{alpha}{The significance level at which cut off predictors. Default is 5\%.}

\item{intercept}{Specify whether the reduction procedure has to consider an 
intercept or not. Default is \code{TRUE}.}
}
\value{
A list with the following elements:\cr
\enumerate{
\item \code{coef}: a matrix of estimated coefficients, standard errors, t-value and
the relative p-values
\item \code{sigma}: the residual standard error and the associated degrees of freedom
\item \code{r.squared}: the R^2 of the model
\item \code{adj.r.squared}: the adjusted R^2 of the model
\item \code{fstatistic}: the F-statistic of the model and the associated degrees of freedom
\item \code{rdf}: the explicit value of the degrees of freedom of the model.
}
}
\description{
The procedure takes a linear model of class \code{lm} (including the output from
a stepwise procedure) and reduces it by iteratively deleting predictors 
based on pvalues of the t-tests. In particular, conversely to what the 
function \code{\link[stats]{lm}} currently does, it allows the inclusion of just 
high order effects. For instance, it can fit just interaction terms given by 
x1*x2 or x1:x2 passed with the usual \code{R} notation.
}
\details{
The procedure makes use of the workhorse function of \code{lm}: 
\code{lm.fit}. This allows to explicitly specify the design matrix which enables
to split main effects from interactions.
}
\examples{
data_factor = mtcars
data_factor$am = as.factor( data_factor$am )
levels( data_factor$am ) = c( "no", "yes" )

complete_model = lm( hp ~ qsec + cyl + mpg + disp + drat + wt + qsec + vs +
                     am + gear + carb + am:qsec, data = data_factor )
summary( complete_model )

reduced_model = reduction( complete_model, data = data_factors )
summary( reduced_model )

}
\author{
Francesco Grossetti \email{francesco.grossetti@gmail.com}.
}
\seealso{
\code{\link[stats]{lm.fit}} \code{\link[stats]{lm}} \code{\link[stats]{step}}
}

