% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statutils.R
\docType{package}
\name{statutils}
\alias{statutils}
\alias{statutils-package}
\title{A collection of tools to facilitate code development.}
\description{
\code{statutils} is a collection of utilities regarding both modelling, 
data management and wrangling. Below is a list
of the current available functions grouped by topic/usage:
\describe{
\item{Linear Models:}{a procedure to reduce a linear model by selecting 
predictors according to a given significance threshold.}
}
}
