% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.list.R
\name{summary.list}
\alias{summary.list}
\title{Summarizing a list as computed by \code{reduction}.}
\usage{
\method{summary}{list}(model, digits = NULL)
}
\arguments{
\item{model}{A \code{list} as computed by \code{\link{reduction}}.}

\item{digits}{An approximate number of digits for printing results. Default is
\code{NULL} so that the function looks to \code{getOption("digits")}.}
}
\value{
A \code{list} with the following arguments:
}
\description{
This function adds a method for \code{\link[base]{summary}} when the object is a list.
In particular, here the list corresponds to a bunch of information computed
through \code{\link[stats]{lm.fit}} and put together by \code{\link{reduction}}.
}
\details{
Bla bla
}
\examples{
data_factor = mtcars
data_factor$am = as.factor( data_factor$am )
levels( data_factor$am ) = c( "no", "yes" )

complete_model = lm( hp ~ qsec + cyl + mpg + disp + drat + wt + qsec + vs +
                     am + gear + carb + am:qsec, data = data_factor )
summary( complete_model )

reduced_model = reduction( complete_model, data = data_factors )
summary( reduced_model )

}
\author{
Francesco Grossetti \email{francesco.grossetti@gmail.com}.
}
\seealso{
\code{\link[stats]{lm.fit}} \code{\link[stats]{lm}} \code{\link[stats]{step}}
}

