% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.list}
\title{Summary method for \code{\link{reduction}}.}
\usage{
summary(model, ...)

\method{summary}{list}(model, digits = NULL)
}
\arguments{
\item{model}{bla bla}
}
\value{
Bla bla
}
\description{
This function adds a method for \code{\link[base]{summary}} when the object is a list.
In particular, here the list corresponds to a bunch of information computed
through \code{\link[stats]{lm.fit}}.
}
\details{
Bla bla
}
\examples{
data_factor = mtcars
data_factor$am = as.factor( data_factor$am )
levels( data_factor$am ) = c( "no", "yes" )

complete_model = lm( hp ~ qsec + cyl + mpg + disp + drat + wt + qsec + vs +
                     am + gear + carb + am:qsec, data = data_factor )
summary( complete_model )

reduced_model = reduction( complete_model, data = data_factors )
summary( reduced_model )

}
\author{
Francesco Grossetti \email{francesco.grossetti@gmail.com}.
}
\seealso{
\code{\link[stats]{lm.fit}} \code{\link[stats]{lm}} \code{\link[stats]{step}}
}

